Handling API Error responses

+++ Reactivities.Application/Activities/Delete.cs
@@ -1,16 +1,17 @@
 ﻿using MediatR;
+using Reactivities.Application.Core;
 using Reactivities.Persistence;
 
 namespace Reactivities.Application.Activities
 {
     public class Delete
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext context;
 
@@ -19,12 +20,19 @@ namespace Reactivities.Application.Activities
                 this.context = context;
             }
 
-            public async Task Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var activity = await context.Activities.FindAsync(request.Id);
 
+                if (activity == null) return null;
+
                 context.Activities.Remove(activity);
-                await context.SaveChangesAsync();
+
+                var result = await context.SaveChangesAsync() > 0;
+
+                if (!result) return Result<Unit>.Failure("Failed to delete activity");
+
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
